+++ lib/services/smart_region_service.dart
@@ -4,7 +4,7 @@ import 'dart:ui' as ui;
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../render/crop_transform.dart';
-import '../render/mask_cache.dart';
+import '../render/brush_rasterizer.dart';
 import '../render/render_engine.dart';
 import '../state/providers.dart';