+++ lib/state/fullscreen_state.dart
@@ -1,3 +1,0 @@
-import 'package:flutter_riverpod/legacy.dart';
-
-final fullscreenPreviewProvider = StateProvider<bool>((ref) => false);