+++ lib/widgets/compare_button.dart
@@ -2,7 +2,7 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../state/compare_state.dart';
+import '../state/params_state.dart';
 
 class CompareButton extends ConsumerWidget {
   final double boxSize;